refactor(backend): colocate profile domain into feature package
Moves the profile controller, entity, repository, DTOs, Caffeine cache
and its cross-replica RabbitMQ invalidation classes out of the generic
controller/cache/dto/database/response layer packages into one
net.onelitefeather.blackhole.backend.profile feature package. Renames
PunishmentProfileHandler to ProfileController to match the naming used
by every other domain controller. Pure physical move, no behavior change.
