+++ backend/src/main/java/net/onelitefeather/blackhole/backend/profile/CacheInvalidationRabbitPublisher.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.cache;
+package net.onelitefeather.blackhole.backend.profile;
 
 import io.micronaut.rabbitmq.annotation.Binding;
 import io.micronaut.rabbitmq.annotation.RabbitClient;
