+++ backend/src/main/java/net/onelitefeather/blackhole/backend/profile/PunishmentProfileRepository.java
@@ -1,10 +1,9 @@
-package net.onelitefeather.blackhole.backend.database.repository;
+package net.onelitefeather.blackhole.backend.profile;
 
 import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.model.Page;
 import io.micronaut.data.model.Pageable;
 import io.micronaut.data.repository.PageableRepository;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentProfileEntity;
 
 @Repository
 public interface PunishmentProfileRepository extends PageableRepository<PunishmentProfileEntity, String> {
