+++ backend/src/main/java/net/onelitefeather/blackhole/backend/punishment/PunishmentApplicationService.java
@@ -2,11 +2,8 @@ package net.onelitefeather.blackhole.backend.punishment;
 
 import io.micronaut.context.BeanProvider;
 import jakarta.inject.Singleton;
-import net.onelitefeather.blackhole.backend.cache.CacheInvalidationPublisher;
 import net.onelitefeather.blackhole.backend.database.entities.PunishmentEntity;
-import net.onelitefeather.blackhole.backend.database.entities.PunishmentProfileEntity;
 import net.onelitefeather.blackhole.backend.database.entities.PunishmentTemplateEntity;
-import net.onelitefeather.blackhole.backend.database.repository.PunishmentProfileRepository;
 import net.onelitefeather.blackhole.backend.database.repository.PunishmentRepository;
 import net.onelitefeather.blackhole.backend.database.repository.PunishmentTemplateRepository;
 import net.onelitefeather.blackhole.backend.dto.EloReasonCode;
@@ -14,6 +11,9 @@ import net.onelitefeather.blackhole.backend.dto.EloTrack;
 import net.onelitefeather.blackhole.backend.dto.PunishType;
 import net.onelitefeather.blackhole.backend.elo.EloService;
 import net.onelitefeather.blackhole.backend.events.DomainEventPublisher;
+import net.onelitefeather.blackhole.backend.profile.CacheInvalidationPublisher;
+import net.onelitefeather.blackhole.backend.profile.PunishmentProfileEntity;
+import net.onelitefeather.blackhole.backend.profile.PunishmentProfileRepository;
 import net.onelitefeather.blackhole.backend.utils.IdGenerator;
 import net.onelitefeather.phoca.metadata.Durationable;
 import net.onelitefeather.phoca.metadata.Expirable;
